chore: mypy cleanup pass — 96 → 0 errors across 38 files
Walked the mypy backlog accumulated across the repo and fixed every
finding, classified by category. mypy now passes clean on all 509
source files. Pyright errors dropped 84 → 31 (remaining are all pre-
existing in modules I didn't touch).

Real bugs found and fixed:

* ``pipeline/bootstrap/pipeline_bootstrap.py:210-217`` — loop variable
  ``e`` collided with the PEP 3110 ``except E as e:`` rule. mypy
  (correctly) flagged "Trying to read deleted variable e" — the for-
  loop's ``e`` got tracked as if it were an except-handler binding.
  Renamed to ``inst_err``. Renamed the parallel loop variable in
  ``cli/commands/plugin.py`` for the same reason.
* ``providers/runpod/pod_control.py`` — added missing ``stop_pod`` /
  ``start_pod`` methods on ``RunPodTrainingPodControl``. The training
  provider's ``pause`` / ``resume`` paths called these methods but the
  control class didn't expose them — they only existed on the
  underlying ``RunPodAPIClient``. So pause/resume on RunPod was broken
  in production at the typed-API layer; mypy caught the gap.
* ``cli/commands/run.py:516`` — ``response.status`` was a typo for the
  actual ``InterruptResponse`` schema (``interrupted`` + ``reason``).
  Replaced with a derived status string.
* ``cli/commands/project.py`` — ``project run`` command called a
  removed ``_exec_orchestrator`` function from ``commands/run.py``;
  ported to the current ``_spawn_worker`` API.
* ``api/services/config_service.py``, ``utils/container.py``,
  ``cli/commands/{config,plugin}.py`` — four call sites still imported
  ``load_config`` from ``src.utils.config`` after that name was renamed
  to ``load_pipeline_config`` and moved to
  ``src.workspace.integrations.loader``. Imports updated; the moved
  symbol is the same one all four call sites already needed.
* ``api/services/connection_test.py`` — ``connect = config.get(...)
  if isinstance(...) else {}`` was typed by mypy as ``dict | None``
  because the ``Any | None`` from ``.get()`` propagated through the
  ternary. Extracted to a local variable so the explicit annotation
  pins ``connect: dict[str, Any]``.

Type narrowing fixes (no runtime behaviour change):

* ``providers/runpod/sdk_adapter.py``, ``pipeline/launch/pod_availability.py``,
  ``providers/runpod/training/provider.py``, ``community/stale_plugins.py``
  — defensive ``isinstance(x, dict)`` guards on values mypy already
  proves narrow are flagged ``unreachable``. Added
  ``# type: ignore[unreachable]`` because the runtime guard remains
  valuable: SDK responses have been observed wrapping the typed shape
  during partial outages.
* ``training/callbacks/system_metrics_callback.py`` — added ``assert
  self._pynvml is not None`` after the ``_pynvml_available`` check so
  mypy can narrow the ``Any | None`` field. Same pattern for psutil.
* ``training/models/loader.py`` — narrowed ``get_adapter_config()``
  return to ``QloraConfig`` via ``isinstance`` assert when
  ``use_4bit`` is true (the only branch that reads bnb_4bit_* fields).
* ``training/mlflow/domain_logger.py`` — same isinstance narrowing for
  the LoraConfig / QloraConfig union; the original code's guard was
  ``config.training.type in ("lora", "qlora")`` which mypy can't use
  to narrow a ``LoraConfig | QloraConfig | AdaLoraConfig`` union.

Missing imports (would break at runtime if the line ever executed):

* ``cli/_smoke_runner.py:407`` — ``threading.Future`` doesn't exist;
  use ``concurrent.futures.Future`` (also dropped the type-arg ignore).
* ``cli/commands/runs.py`` — ``Any`` annotation used without import.
* ``api/services/control_plane_heartbeat.py:103`` — ``"callable | None"``
  string literal for a parameter type uses the builtin ``callable``
  function as a type, which is invalid. Replaced with
  ``Callable[..., object] | None`` from collections.abc.

Bare-comment / em-dash type-comment hazards:

* ``config/training/schema.py`` — three lines like ``#   type: lora    →
  lora: ...`` matched mypy's ``# type:`` comment-directive parser. The
  em-dash is invalid Python syntax inside a type-comment, so mypy's
  parsing of the entire import chain choked on this line. Reformatted
  to ``#   - "lora"    -> lora: ...`` so mypy doesn't interpret it.

Misc:

* ``providers/single_node/inference/artifacts.py`` —
  ``render_readme(endpoint_url: str)`` widened to ``str | None``
  matching the runpod side; placeholder string used when None. Caller
  in ``provider.py`` was passing ``str | None`` from ``EndpointInfo``.
* ``providers/single_node/training/provider.py`` —
  ``cleanup_after_run(ssh_command_timeout: float)`` reverted to ``int``
  because the underlying ``SSHClient.exec_command`` requires int and
  the only call site passes 10.
* ``community/loader.py`` — ``set(dict)`` over an ``Any``-keyed
  mapping confused mypy's set-difference inference; replaced with
  explicit ``{str(k) for k in ...}`` set-comprehensions. Cast on
  ``ALL_PLUGIN_KINDS`` element to satisfy the ``PluginKind`` Literal
  parameter of ``load_plugins``.
* Multiple files: callable types tightened from ``object | None`` to
  ``Callable[..., Any] | None`` so direct invocation is type-clean.
* ``pipeline/launch/runtime.py`` — explicit ``process: subprocess.Popen[Any]``
  annotation so the foreground/detached branches (Popen[bytes] vs
  Popen[str] with text=True) reconcile.
* ``training/managers/mlflow_manager/{setup,manager}.py`` — multiple
  ``# type: ignore[attr-defined,assignment,unreachable]`` markers
  added where the mixin's narrower-than-concrete declarations fight
  the concrete class's runtime types; the runtime path is correct,
  this is purely about telling mypy to stop complaining about the
  mixin-class declaration mismatch. Refactoring the mixin into a
  proper Protocol is a larger structural change out of scope here.

Pre-commit ran ruff format on the touched files; the diff is wider
than the actual logic changes for that reason.

Per-hook status on touched files:

* ruff-format: PASS
* ruff-lint: 32 remaining findings are all pre-existing (ARG002 on HF
  Trainer callback signatures with mandatory args, E402 on the
  intentional late import in providers/runpod/training/provider.py,
  TC001/TC002/TC003 on import locations in modules I didn't author).
* mypy: PASS (96 → 0)
* pyright: PASS for all touched files (84 → 31, the 31 remaining are
  in pre-existing modules I didn't touch).
* bandit: PASS

Targeted test regression on touched files: 761 passing, 1 skipped.
The full repo unit suite has 214 pre-existing failures across modules
unrelated to this work — confirmed by running the same query on a
clean stash. Out of scope for this commit.

+++ src/providers/single_node/training/provider.py
@@ -326,7 +326,7 @@ class SingleNodeProvider(IGPUProvider):
         self,
         container_name: str,
         *,
-        ssh_command_timeout: float = 10.0,
+        ssh_command_timeout: int = 10,
     ) -> Result[None, ProviderError]:
         """Phase 9.B — terminate the training docker container, fail-soft.
 
@@ -370,10 +370,7 @@ class SingleNodeProvider(IGPUProvider):
         if self._ssh_client is None:
             return Err(
                 ProviderError(
-                    message=(
-                        "cleanup_after_run called before connect — "
-                        "no SSH client to drive docker rm"
-                    ),
+                    message=("cleanup_after_run called before connect — " "no SSH client to drive docker rm"),
                     code="SINGLENODE_CLEANUP_NO_SSH",
                 )
             )
@@ -384,9 +381,7 @@ class SingleNodeProvider(IGPUProvider):
         # We don't pass the value through a shell — exec_command
         # handles quoting — but explicit narrow allowlist is
         # cleaner and surfaces typos early.
-        if not container_name or any(
-            ch in container_name for ch in (" ", ";", "&", "|", "$", "`", "\n")
-        ):
+        if not container_name or any(ch in container_name for ch in (" ", ";", "&", "|", "$", "`", "\n")):
             return Err(
                 ProviderError(
                     message=(
@@ -403,14 +398,14 @@ class SingleNodeProvider(IGPUProvider):
         rm_cmd = f"docker rm -f {container_name} >/dev/null 2>&1 || true"
         try:
             ok, _stdout, stderr = self._ssh_client.exec_command(
-                command=rm_cmd, timeout=ssh_command_timeout, silent=False,
+                command=rm_cmd,
+                timeout=ssh_command_timeout,
+                silent=False,
             )
-        except Exception as exc:  # noqa: BLE001 — SSH may raise generic
+        except Exception as exc:
             return Err(
                 ProviderError(
-                    message=(
-                        f"cleanup_after_run: SSH transport failed: {exc}"
-                    ),
+                    message=(f"cleanup_after_run: SSH transport failed: {exc}"),
                     code="SINGLENODE_CLEANUP_SSH_TRANSPORT",
                 )
             )
@@ -418,10 +413,7 @@ class SingleNodeProvider(IGPUProvider):
         if not ok:
             return Err(
                 ProviderError(
-                    message=(
-                        f"docker rm -f {container_name} failed: "
-                        f"{(stderr or '')[:200]}"
-                    ),
+                    message=(f"docker rm -f {container_name} failed: " f"{(stderr or '')[:200]}"),
                     code="SINGLENODE_CLEANUP_DOCKER_RM_FAILED",
                 )
             )
@@ -431,15 +423,17 @@ class SingleNodeProvider(IGPUProvider):
         verify_cmd = f"docker ps -a -q -f name=^{container_name}$"
         try:
             ok_v, stdout_v, _ = self._ssh_client.exec_command(
-                command=verify_cmd, timeout=ssh_command_timeout, silent=True,
+                command=verify_cmd,
+                timeout=ssh_command_timeout,
+                silent=True,
             )
-        except Exception as exc:  # noqa: BLE001
+        except Exception as exc:
             # Verification failed but rm succeeded — log and treat as
             # success; the rm output is the source of truth.
             logger.debug(
-                "[PROVIDER:CLEANUP] verify step failed for %s: %s — "
-                "treating as removed since docker rm succeeded",
-                container_name, exc,
+                "[PROVIDER:CLEANUP] verify step failed for %s: %s — " "treating as removed since docker rm succeeded",
+                container_name,
+                exc,
             )
             return Ok(None)
 
@@ -447,15 +441,15 @@ class SingleNodeProvider(IGPUProvider):
             return Err(
                 ProviderError(
                     message=(
-                        f"docker rm reported success but container "
-                        f"{container_name!r} still listed by docker ps"
+                        f"docker rm reported success but container " f"{container_name!r} still listed by docker ps"
                     ),
                     code="SINGLENODE_CLEANUP_VERIFY_FAILED",
                 )
             )
 
         logger.info(
-            "[PROVIDER:CLEANUP] training container %s removed", container_name,
+            "[PROVIDER:CLEANUP] training container %s removed",
+            container_name,
         )
         return Ok(None)
 
@@ -598,8 +592,8 @@ class SingleNodeProvider(IGPUProvider):
             has_pause_resume=False,
             runner_workspace_root="/workspace",
             # Phase 14.D+F capability fields:
-            is_local=True,                  # always-on host
-            supports_log_download=False,    # logs already on host filesystem
+            is_local=True,  # always-on host
+            supports_log_download=False,  # logs already on host filesystem
         )
 
     def required_secrets(self) -> tuple[str, ...]:
@@ -620,7 +614,9 @@ class SingleNodeProvider(IGPUProvider):
     # ------------------------------------------------------------------
 
     def required_runtime_env_vars(
-        self, *, resource_id: str | None,
+        self,
+        *,
+        resource_id: str | None,
     ) -> dict[str, str]:
         """Env vars the in-pod runner needs.
 
@@ -636,7 +632,8 @@ class SingleNodeProvider(IGPUProvider):
         return {RUNTIME_PROVIDER_ENV_VAR: PROVIDER_SINGLE_NODE}
 
     def probe_availability(
-        self, resource_id: str,
+        self,
+        resource_id: str,
     ) -> AvailabilityVerdict:
         """Single_node host is always-on — return ``running`` immediately.
